manager.py = the manager for the tracking and the landmark, if a user doesn't give a good video for the LPC process. coordswriter.py, manageroptions.py, manageroutputs.py = the writer, option and outputs file which work with manager.
managerdetection.py = the manager for the detection process on an image.
detectionwriter.py, detectionoptions.py, detectionoutputs.py = the writer, option and outputs file which work with managerdetection.

trackingmanager.py = the manager for the tracking process which use face detection.
trackingwriter.py, trackingoptions.py, trackingoutputs.py = the writer, option and outputs file which work with trackingmanager.

landmarkmanager.py = the manager for the detection process on an image.
landmarkwriter.py, landmarkoptions.py, landmarkoutputs.py = the writer, option and outputs file which work with landmarkmanager.

landmarkoutputs.py, manageroutputs.py, trackingoutputs.py = each of them contains a class which is child of VideoOutputs.

videooutputs.py = contains the mother class VideoOutputs which is the super class of LandmarkOutputs, TrackingOutput, ManagerOutputs.